% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_modified_logistic.R
\name{fit.modified.logistic}
\alias{fit.modified.logistic}
\title{Fitting Modified Logistic}
\usage{
fit.modified.logistic(
  I.vec,
  t.vec,
  K.start = 1e+05,
  theta.start = 0.1,
  omega.start = 0.1,
  gamma.start = 1,
  autoAdjustStartingVals = TRUE,
  autoAdjustedTheta.start = 1e-06,
  adjustmentOmega.start = -0.1,
  adjustmentGamma.start = 1,
  adjustmentK.start = 1e+05,
  maxiter = 10000,
  tol = 1e-04,
  minFactor = 1e-15,
  adjustForMinCutoff = TRUE,
  minCutoff = 0,
  adjust_t_ForMinCutoff = TRUE,
  constFor_t_Adjustment = 1
)
}
\arguments{
\item{I.vec}{Vector of values of the output variable - number of infected at given point in time in case of epidemic data.}

\item{t.vec}{Vector of values of the input variable - value of time in case of time series.}

\item{K.start}{Starting value of parameter K given to the optimization algorithm which adjusts the coefficients.}

\item{theta.start}{Starting value of parameter theta given to the optimization algorithm which adjusts the coefficients.}

\item{omega.start}{Starting value of parameter omega given to the optimization algorithm which adjusts the coefficients.}

\item{gamma.start}{Starting value of parameter gamma given to the optimization algorithm which adjusts the coefficients.}

\item{autoAdjustStartingVals}{IF TRUE the algorithm will first fit the model without the theta term and then use the obtained values as starting values for fitting of the full model. It is recommended to leave this TRUE.}

\item{autoAdjustedTheta.start}{Starting value of parameter theta in case the model is fitted using the other parameters from the reduced model fitting. It is recommended to keep this one close to 0.}

\item{adjustmentOmega.start}{Starting value of parameter omega given to the optimization algorithm which adjusts the coefficients in the reduced model.}

\item{adjustmentGamma.start}{Starting value of parameter gamma given to the optimization algorithm which adjusts the coefficients in the reduced model.}

\item{adjustmentK.start}{Starting value of parameter K given to the optimization algorithm which adjusts the coefficients in the reduced model.}

\item{maxiter}{A positive integer specifying the maximum number of iterations allowed for both reduced and full model fitting.}

\item{tol}{A positive numeric value specifying the tolerance level for the relative offset convergence criterion for both reduced and full model fitting.}

\item{minFactor}{A positive numeric value specifying the minimum step-size factor allowed on any step in the iteration.}

\item{adjustForMinCutoff}{If TRUE the function will delete the data points for which the value of I.vec is less than the minCutoff.}

\item{minCutoff}{If adjustForMinCutoff is TRUE the function will delete the data points for which the value of I.vec is less than the minCutoff.}

\item{adjust_t_ForMinCutoff}{If the function adjusted for the minCutoff and this parameter is TRUE, the valule of t will be adjusted so that t starts running from the first data point which was not deleted.}

\item{constFor_t_Adjustment}{This value will be added after the adjustment to t is made post miCutoff adjustment. It is recommended to keep this parameter at value greater than 0, so that t is never 0.}
}
\value{
Nonlinear regression model.
}
\description{
fit.modified.logistic is used to fit the modified logistic model as specified in the description of the package.
}
\examples{
fit.modified.logistic(I.vec = data$Cases, t.vec = data$Days, K.start = 200000, omega.start = 0.2, gamma.start = 2, autoAdjustedTheta.start = 1e-7, minCutoff = 500)
}
