% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modified_logistic.R
\name{plot.modified.logistic}
\alias{plot.modified.logistic}
\title{Plot the Fitted Model Anlong with Data Points}
\usage{
\method{plot}{modified.logistic}(model, I.vec, t.vec, numIntervals = 1e+05, xlim, ylim)
}
\arguments{
\item{model}{Model to be plotted.}

\item{I.vec}{Vector of y-coordinates of data points to be plotted.}

\item{t.vec}{Vector of x-coordinates of data points to be plotted.}

\item{numIntervals}{Number of intervals into which we divide the scope of the plot to provide for a smoothly-looking curve. It is recommended to keep this one large.}

\item{xlim}{Vector with two values - lower and upper limit for the x-coordinate in the scope of the plot.}

\item{ylimVector}{Vector with two values - lower and upper limit for the y-coordinate in the scope of the plot.}
}
\value{

}
\description{
plot.modified.logistic is used to plot the modified logistic model as specified in the description of the package together with the data points which were used to fit it.
}
\examples{
plot.modified.logistic(model, data$Cases, data$Days, xlim = c(1, 200), ylim = c(0, 1000000))
}
